fix: rename open type from formatter to transform
