+++ index.js
@@ -2,7 +2,7 @@ const JsonSetsFormatter = require("./lib/json-sets-formatter");
 const AttributeSetsTransform = require("./lib/attribute-sets-transform");
 const NameKebabTransfom = require("./lib/name-kebab-transform");
 const CSSSetsFormatter = require("./lib/css-sets-formatter.js");
-const CSSOpenTypeFormatter = require("./lib/css-font-open-type-formatter.js");
+const CSSOpenTypeTransform = require("./lib/css-font-open-type-transform.js");
 const DroverJsonFormatter = require("./lib/drover-json-formatter.js");
 
 module.exports = {
@@ -10,6 +10,6 @@ module.exports = {
   AttributeSetsTransform,
   NameKebabTransfom,
   CSSSetsFormatter,
-  CSSOpenTypeFormatter,
+  CSSOpenTypeTransform,
   DroverJsonFormatter,
 };
